Trim comments and fold the preference tests into one file
